+++ raiden/network/transport/__init__.py
@@ -1,0 +1,2 @@
+from raiden.network.transport.matrix import MatrixTransport  # NOQA
+from raiden.network.transport.udp import UDPTransport        # NOQA
