+++ raiden/network/transport/udp/__init__.py
@@ -1,0 +1,1 @@
+from raiden.network.transport.udp.udp_transport import UDPTransport  # NOQA
